fix: correct audit id bigint mapping
- Change AuditId from int to long
- Read audit Id using GetInt64
- Update SQL Id parameters to BIGINT
- Use SqlDbType.BigInt for audit status procedures
- Align stored procedures with BIGINT audit identifiers

+++ Models/HospitalizationAuditEvent.cs
@@ -2,7 +2,7 @@ namespace HospitalizationReconciliationNewRelic.Models;
 
 public sealed class HospitalizationAuditEvent
 {
-    public int AuditId { get; set; }
+    public long AuditId { get; set; }
 
     public string? Container { get; set; }
